+++ ui/src/components/avatar/QasAvatar.vue
@@ -1,5 +1,5 @@
 <template>
-  <q-avatar :class="classes" :color="backgroundColor" v-bind="$attrs" v-on="$listeners">
+  <q-avatar :class="classes" rounded v-bind="$attrs" v-on="$listeners">
     <q-img v-if="hasImage" :alt="title" :ratio="1" spinner-color="primary" spinner-size="16px" :src="image" @error="onImageLoadedError" />
     <template v-else-if="hasTitle">{{ firstLetter }}</template>
     <q-icon v-else :name="icon" />
@@ -7,32 +7,14 @@
 </template>
 
 <script>
-export const colors = [
-  'amber',
-  'blue',
-  'blue-grey',
-  'brown',
-  'cyan',
-  'deep-orange',
-  'deep-purple',
-  'green',
-  'grey',
-  'indigo',
-  'light-blue',
-  'light-green',
-  'lime',
-  'orange',
-  'pink',
-  'purple',
-  'red',
-  'teal',
-  'yellow'
-]
-
 export default {
   props: {
+    color: {
+      default: 'primary',
+      type: String
+    },
+
     dark: {
-      default: false,
       type: Boolean
     },
 
@@ -54,11 +36,6 @@ export default {
     title: {
       default: '',
       type: String
-    },
-
-    token: {
-      default: 0,
-      type: Number
     }
   },
 
@@ -69,19 +46,18 @@ export default {
   },
 
   computed: {
-    backgroundColor () {
-      if (this.color) {
-        return this.color
-      }
-      return this.dark ? this.mainColor : `${this.mainColor}-2`
-    },
-
     classes () {
-      return [this.textColorClass]
+      const contrastColor = this.textColor ? this.textColor : this.contrastColor
+
+      return [
+        this.dark
+          ? `bg-${this.color} text-${contrastColor}`
+          : `bg-${contrastColor} text-${this.color}`
+      ]
     },
 
-    color () {
-      return this.$attrs.color
+    contrastColor () {
+      return `${this.color}-contrast`
     },
 
     firstLetter () {
@@ -92,31 +68,8 @@ export default {
       return !this.hasImageError && !!this.image
     },
 
-    hasTextColor () {
-      return !!this.textColor
-    },
-
     hasTitle () {
       return !!this.title
-    },
-
-    mainColor () {
-      if (!this.title && !this.token) {
-        return 'grey'
-      }
-
-      const dividend = this.title ? this.title.length : this.token
-      return colors[dividend % colors.length]
-    },
-
-    textColorClass () {
-      if (this.hasTextColor) {
-        return `text-${this.textColor}`
-      }
-
-      return (this.color === 'black' || this.dark)
-        ? `text-${this.mainColor}-2`
-        : `text-${this.mainColor}`
     }
   },